backend/requirements.txt
psycopg2
django==4.1.1
djangorestframework~=3.12.4
requests
django-dotenv
django-cors-headers
pytz
